Program for Matching Pairs Game
I'm learning how to combine actions of a form with a Program. I will have either convert my .vb form to a .cs form, or recreate the form.

Once I fix my form, the Program will generate random locations for the 8 pairs of icons that I assign to the form (which is a 4*4 grid). Then, the user will be able to play the game.

The color of the icon becomes of the color of the entire square when a match is selected.