Add files via upload
This is a result of a short Udemy tutorial from Bluelime Learning Solutions. I created an embedded web browser (internet explorer). It's a bit buggy, as it uses an outdated version of Internet Explorer, but it is fully functional. The logic used is in the Form1.cs file.

I plan to see how other tools can be used to create a web browser.